Uso mais eficiente da memória no PID.
Uso mais eficiente da memória mudando alguns tipos de dados de int para uint8_t ou uint16_t.

+++ Controle/src/PID.h
@@ -15,6 +15,8 @@
 
 #ifndef PID_h
 #define PID_h
+
+#include "type.h"
 //TODO: verificar utilidade desse define no projeto
 //#define LIBRARY_VERSION	1.0.0
 
@@ -34,32 +36,32 @@ typedef struct pid_data_t {
     float dispKp;				   // * we'll hold on to the tuning parameters in user-entered
     float dispKi;				   //   format for display purposes
     float dispKd;				   //
-    int   controllerDirection;
+    uint8_t controllerDirection;
     float ITerm, lastInput;
-    int   SampleTime;
+    uint16_t SampleTime;
     float outMin, outMax;
-    int   inAuto;
+    uint8_t   inAuto;
 }* pid_data_t;
 
 //commonly used functions **************************************************************************
 void pid_update_data(pid_data_t, float Input, float Setpoint);
-void pid_setMode(pid_data_t, int);                   // * sets PID to either PID_MANUAL (0) or PID_AUTOMATIC (non-0)
+void pid_setMode(pid_data_t, uint8_t);               // * sets PID to either PID_MANUAL (0) or PID_AUTOMATIC (non-0)
 void pid_compute(pid_data_t pid_data);               // * performs the PID calculation.
 void pid_setOutputLimits(pid_data_t, float, float);  // * sets the output to a specific range. 0-255 by default
 
 
 //available but not commonly used functions ********************************************************
 void pid_setTunings(pid_data_t,float, float, float);    // pid_data, Kp, Ki, Kd
-void pid_setControllerDirection(pid_data_t, int);	    // pid_data, PID_DIRECT ou PID_REVERSE
-void pid_setSampleTime(pid_data_t, int);                // pid_data, NewSampleTime
+void pid_setControllerDirection(pid_data_t, uint8_t);	// pid_data, PID_DIRECT ou PID_REVERSE
+void pid_setSampleTime(pid_data_t, uint16_t);           // pid_data, NewSampleTime
 void pid_initialize(pid_data_t);
 
 //Display functions ****************************************************************
 float pid_getKp(pid_data_t);						  // These functions query the pid for interal values.
 float pid_getKi(pid_data_t);						  //  they were created mainly for the pid front-end,
 float pid_getKd(pid_data_t);						  // where it's important to know what is actually
-int pid_getMode(pid_data_t);						  //  inside the PID.
-int pid_getDirection(pid_data_t);					  //
+uint8_t pid_getMode(pid_data_t);						  //  inside the PID.
+uint8_t pid_getDirection(pid_data_t);					  //
 
 #endif
